+++ editor/editor-plugin-type-ahead/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-plugin-type-ahead
 
+## 13.0.2
+
+### Patch Changes
+
+- [`f8280f6b24eec`](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/commits/f8280f6b24eec) -
+  Allow editor surfaces to control agent mention picker sectioning independently.
+
 ## 13.0.1
 
 ### Patch Changes